app_kbqa/BAMnet/requirements.txt
pytorch==0.4.1
